# Speed up ResultMatcher implict lookups