--- a/core/src/main/scala/org/http4s/rho/package.scala
+++ b/core/src/main/scala/org/http4s/rho/package.scala
@@ -20,7 +20,8 @@ trait RhoDsl[F[_]]
     with QueryParsers[F]
     with MatchersHListToFunc[F]
     with ResponseGeneratorInstances[F]
-    with FailureResponseOps[F] {
+    with FailureResponseOps[F]
+    with ResultMatchers[F] {
 
   implicit def method(m: Method): PathBuilder[F, HNil] = new PathBuilder(m, PathEmpty)