This is a follow up to the profiling of compilation times of rho routes started in #417.

Let's start with an analysis of the current performance profile of the [Benchmark5_TypedPathAndComplexResult](https://github.com/RafalSumislawski/rho/blob/compilation-benchmarks/benchmark/src/main/scala/com/http4s/rho/swagger/demo/Benchmark5_TypedPathAndComplexResult.scala). 

![before-with-annotations](https://user-images.githubusercontent.com/9465998/82737964-ba703f00-9d34-11ea-8637-29a54f541b2e.png)

The whole width of the flame chart is 12640ms. The lookup for ResultMatcher[F, F[Result[...]]] that I'm trying to optimize in this PR is 9450ms (75%).

- A - failed attempts to use the `writableMatcher` - problem solved by 
d5de0a6 
- B - `FnToProduct` and a `Reverse` this part will be doubled by #417
- C - building HList of the inputs (path vars in this case)
- D - lots of searches for `maybeWritableAny` (on for each `Nothing` in `Result`) - these are made faster by 0024904

d5de0a6 is a general improvement of implicit lookups by fixing the `F`

After these changes the `Benchmark5_TypedPathAndComplexResult` looks like this:

![after](https://user-images.githubusercontent.com/9465998/82737813-e50dc800-9d33-11ea-8337-babd5bcb37dc.PNG)
The whole chart is 8006ms and the `ResultMatcher` lookup is 4478m (56%). So the changes **reduced compilation time by ~37%.**



### Commit by commit measurements
The three commits are to some extent independent, so this PR isn't all-or-nothing, we can decide to merge some of them while leaving out others.

These are measurement of the `Benchmark7_ResultMatcherFOk` - the benchmark that measures just the ResultMatcher lookup:
- 9341.225ms // master
- 8076.063ms // d5de0a6  writable as low prio
- 6033.998ms // 0024904  remove MaybeWritable.Aux
- 4826.143ms // bce27dd  object -> trait typed with F